+++ app/public/locales/ja/translation.json
@@ -781,6 +781,7 @@
     "import": {
         "scan": "QRスキャン",
         "manual": "手動入力",
+        "passkey": "パスキー(ベータ)",
         "getStarted": "はじめる",
         "or": "または",
         "qrHint": "ログイン用のQRコードは、ログイン済みのデバイスの「設定 > ログインQR」から取得できます。",
