+++ app/src/pages/AccountImport.tsx
@@ -1,120 +1,26 @@
-import { Alert, AlertTitle, Box, Button, Divider, Paper, Typography } from '@mui/material'
+import { Alert, Box, Button, Divider, Paper, Typography } from '@mui/material'
 import { GuestBase } from '../components/GuestBase'
 import { ImportMasterKey } from '../components/Importer/ImportMasterkey'
 import { Link } from 'react-router-dom'
 import { ImportSubkey } from '../components/Importer/ImportSubkey'
 
-import QrCodeScannerIcon from '@mui/icons-material/QrCodeScanner'
-import PasswordIcon from '@mui/icons-material/Password'
 import { IconButtonWithLabel } from '../components/ui/IconButtonWithLabel'
-import { Trans, useTranslation } from 'react-i18next'
-import { Suspense, lazy, useEffect, useState } from 'react'
+import { useTranslation } from 'react-i18next'
+import { Suspense, lazy, useState } from 'react'
 import { Client } from '@concrnt/worldlib'
 import { Helmet } from 'react-helmet-async'
-import { DeriveIdentity, LoadSubKey } from '@concrnt/client'
-import { string2Uint8Array } from '../util'
-import { enqueueSnackbar } from 'notistack'
+
+import QrCodeScannerIcon from '@mui/icons-material/QrCodeScanner'
+import PasswordIcon from '@mui/icons-material/Password'
+import VpnKeyIcon from '@mui/icons-material/VpnKey'
+import { ImportPasskey } from '../components/Importer/ImportPasskey'
 
 const QRCodeReader = lazy(() => import('../components/ui/QRCodeReader'))
-
-function bufToHex(ab: ArrayBuffer): string {
-    return Array.from(new Uint8Array(ab))
-        .map((b) => b.toString(16).padStart(2, '0'))
-        .join('')
-}
 
 export default function AccountImport(): JSX.Element {
     const { t } = useTranslation('', { keyPrefix: 'import' })
 
-    const [importMode, setImportMode] = useState<'none' | 'scan' | 'manual'>('none')
-
-    useEffect(() => {
-        // receive passkey
-        const run = async () => {
-            const challenge = new Uint8Array(32)
-            crypto.getRandomValues(challenge)
-            const cred = await navigator.credentials.get({
-                publicKey: {
-                    challenge: challenge,
-                    rpId: window.location.hostname,
-                    userVerification: 'required',
-                    extensions: {
-                        prf: {
-                            eval: {
-                                first: string2Uint8Array('concrnt-world-passkey')
-                            }
-                        }
-                    }
-                }
-            })
-
-            console.log('cred', cred)
-            if (!cred) {
-                console.error('Failed to get passkey')
-                enqueueSnackbar('Failed to get passkey.', { variant: 'error' })
-                return
-            }
-
-            // @ts-ignore
-            const userHandle = cred.response?.userHandle
-            if (!userHandle) {
-                console.error('No user handle found in passkey response')
-                enqueueSnackbar('No user handle found in passkey response.', { variant: 'error' })
-                return
-            }
-
-            const domain = new TextDecoder().decode(userHandle)
-
-            // @ts-ignore
-            const credentialResults = cred.getClientExtensionResults()
-            console.log('Credential Results:', credentialResults)
-            const prfRes = credentialResults?.prf?.results
-            if (prfRes?.first) {
-                console.log('PRF First:', prfRes.first)
-                const firstBuf = new Uint8Array(prfRes.first)
-                console.log('source:', bufToHex(prfRes.first))
-
-                const identity = DeriveIdentity(firstBuf)
-                console.log('Derived Identity:', identity)
-
-                const dummySubkeyStr = `concurrent-subkey ${identity.privateKey} -@${domain} -`
-                const dummySubkey = LoadSubKey(dummySubkeyStr)
-
-                if (!dummySubkey) {
-                    console.error('Failed to load dummy subkey')
-                    enqueueSnackbar('Failed to load dummy subkey.', { variant: 'error' })
-                    return
-                }
-
-                const res = await fetch(`https://${domain}/api/v1/key/${dummySubkey.ckid}`, {})
-                    .then((response) => response.json())
-                    .catch((error) => {
-                        console.error('Failed to fetch key:', error)
-                        enqueueSnackbar('Failed to fetch key.', { variant: 'error' })
-                        return null
-                    })
-
-                if (!res || !res.content || res.content.length === 0) {
-                    console.error('Invalid response from server:', res)
-                    enqueueSnackbar('Account not found.', { variant: 'error' })
-                    return
-                }
-
-                const ccid = res.content[0].root
-                const subkeyStr = `concurrent-subkey ${identity.privateKey} ${ccid}@${domain} -`
-
-                localStorage.setItem('Domain', JSON.stringify(domain))
-                localStorage.setItem('SubKey', JSON.stringify(subkeyStr))
-                window.location.href = '/'
-            } else {
-                console.error('No PRF first result found')
-                enqueueSnackbar('Provided passkey is not supported.', { variant: 'error' })
-                return
-            }
-        }
-
-        run()
-    }, [])
+    const [importMode, setImportMode] = useState<'none' | 'scan' | 'passkey' | 'manual'>('none')
 
     return (
         <GuestBase
@@ -162,6 +68,14 @@ export default function AccountImport(): JSX.Element {
                     />
                     <Divider orientation="vertical">{t('or')}</Divider>
                     <IconButtonWithLabel
+                        icon={VpnKeyIcon}
+                        label={t('passkey')}
+                        onClick={() => {
+                            setImportMode('passkey')
+                        }}
+                    />
+                    <Divider orientation="vertical">{t('or')}</Divider>
+                    <IconButtonWithLabel
                         icon={PasswordIcon}
                         label={t('manual')}
                         onClick={() => {
@@ -197,19 +111,7 @@ export default function AccountImport(): JSX.Element {
                         <ImportSubkey />
                     </>
                 )}
-                {importMode === 'none' && (
-                    <>
-                        <Alert severity="info">
-                            <AlertTitle>{t('passkeyNoticeTitle')}</AlertTitle>
-                            {t('passkeyNoticeBody')}
-                        </Alert>
-
-                        <Alert severity="warning">
-                            <AlertTitle>{t('passkeyWarnTitle')}</AlertTitle>
-                            <Trans i18nKey="import.passkeyWarnBody" />
-                        </Alert>
-                    </>
-                )}
+                {importMode === 'passkey' && <ImportPasskey />}
             </Paper>
         </GuestBase>
     )
